requirements.txt
PyPDF2==2.12.1
camelot-py
"camelot-py[cv]"
unidecode
pyodbc